+++ lib/output/broker_out_common.go
@@ -26,6 +26,8 @@ import (
 	"strconv"
 	"strings"
 
+	"github.com/Jeffail/benthos/lib/broker"
+
 	"gopkg.in/yaml.v2"
 )
 
@@ -82,58 +84,35 @@ func (b *brokerOutputList) UnmarshalYAML(unmarshal func(interface{}) error) erro
 // formats that we do not know at this stage (JSON, YAML, etc), therefore we use
 // the more hacky method as performance is not an issue at this stage.
 func parseOutputConfsWithDefaults(outConfs []interface{}) ([]Config, error) {
-	type confAlias Config
-
 	outputConfs := []Config{}
 
 	for i, boxedConfig := range outConfs {
-		newConfs := []confAlias{confAlias(NewConfig())}
-		if i > 0 {
-			// If the type of this output is 'ditto' we want to start with a
-			// duplicate of the previous config.
-			newConfsFromDitto := func(label string) error {
-				// Remove the vanilla config.
-				newConfs = []confAlias{}
+		newConfs := make([]Config, 1)
+		label := broker.GetGenericType(boxedConfig)
 
-				if len(label) > 5 && label[5] == '_' {
-					if label[6:] == "0" {
-						// This is a special case where we are expressing that
-						// we want to end up with zero duplicates.
-						return nil
-					}
+		if i > 0 && strings.Index(label, "ditto") == 0 {
+			broker.RemoveGenericType(boxedConfig)
+
+			// Check if there is a ditto multiplier.
+			if len(label) > 5 && label[5] == '_' {
+				if label[6:] == "0" {
+					// This is a special case where we are expressing that
+					// we want to end up with zero duplicates.
+					newConfs = nil
+				} else {
 					n, err := strconv.Atoi(label[6:])
 					if err != nil {
-						return fmt.Errorf("failed to parse ditto multiplier: %v", err)
+						return nil, fmt.Errorf("failed to parse ditto multiplier: %v", err)
 					}
-					for j := 0; j < n; j++ {
-						newConfs = append(newConfs, confAlias(outputConfs[i-1]))
-					}
-				} else {
-					newConfs = append(newConfs, confAlias(outputConfs[i-1]))
+					newConfs = make([]Config, n)
 				}
-				return nil
+			} else {
+				newConfs = make([]Config, 1)
 			}
-			switch unboxed := boxedConfig.(type) {
-			case map[string]interface{}:
-				if t, ok := unboxed["type"].(string); ok && strings.Index(t, "ditto") == 0 {
-					if err := newConfsFromDitto(t); err != nil {
-						return nil, err
-					}
-					if len(newConfs) > 0 {
-						unboxed["type"] = newConfs[0].Type
-					}
-				}
-			case map[interface{}]interface{}:
-				if t, ok := unboxed["type"].(string); ok && strings.Index(t, "ditto") == 0 {
-					if err := newConfsFromDitto(t); err != nil {
-						return nil, err
-					}
-					if len(newConfs) > 0 {
-						unboxed["type"] = newConfs[0].Type
-					}
-				}
-			}
+
+			broker.ComplementGenericConfig(boxedConfig, outConfs[i-1])
 		}
+
 		for _, conf := range newConfs {
 			rawBytes, err := yaml.Marshal(boxedConfig)
 			if err != nil {
@@ -142,7 +121,7 @@ func parseOutputConfsWithDefaults(outConfs []interface{}) ([]Config, error) {
 			if err = yaml.Unmarshal(rawBytes, &conf); err != nil {
 				return nil, err
 			}
-			outputConfs = append(outputConfs, Config(conf))
+			outputConfs = append(outputConfs, conf)
 		}
 	}